requirements.txt
fastapi==0.109.2
fastapi-mail
pydantic==2.6.1
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
sqlalchemy==2.0.27
alembic==1.13.1
gunicorn==21.2.0
pytest==7.4.2
httpx==0.25.0
coverage>=6.0.2,<6.1.0
faker>=9.3.1,<9.4.0
boto3==1.28.36
stripe==6.5.0
jinja2>=3.0.2,<3.1.0
redis==5.0.1
celery==5.3.4
flower>=1.0.0,<1.1.0
sentry-sdk>=1.4.3,<1.5.0
unidecode==1.3.7
python-dotenv==1.0.1
email-validator==2.0.0.post2
pyjwt==2.8.0
pydantic-settings==2.1.0
bcrypt==4.1.2
psycopg2-binary==2.9.9

# AI Image Generation
pillow==10.2.0
aiofiles==23.2.1
asyncio==3.4.3

# AWS S3 and Local Storage
python-magic==0.4.27
aiohttp==3.9.1
# aiobotocore==2.11.0

# Development and Testing
pytest-asyncio==0.23.2
black==23.12.0
isort==5.13.2
flake8==7.0.0

# Monitoring and Logging
structlog==23.2.0
